refactor(lightbox): move pager files to features/lightbox/pager
