+++ src/components/Post/Embed/ImageEmbed.tsx
@@ -2,7 +2,6 @@ import {InteractionManager, View} from 'react-native'
 import {type AnimatedRef} from 'react-native-reanimated'
 import {Image} from 'expo-image'
 
-import {type Dimensions} from '#/view/com/lightbox/ImageViewing/@types'
 import {atoms as a, tokens} from '#/alf'
 import {AutoSizedImage} from '#/components/images/AutoSizedImage'
 import {Gallery} from '#/components/images/Gallery'
@@ -10,6 +9,7 @@ import {ImageLayoutGrid} from '#/components/images/ImageLayoutGrid'
 import {PostEmbedViewContext} from '#/components/Post/Embed/types'
 import {useAnalytics} from '#/analytics'
 import {useLightboxControls} from '#/features/lightbox/state'
+import {type Dimensions} from '#/features/lightbox/types'
 import {type EmbedType} from '#/types/bsky/post'
 import {type CommonProps} from './types'
