+++ src/features/lightbox/types.ts
